requirements.txt
certifi==2017.4.17
chardet==3.0.4
click==6.7
idna==2.5
nltk==3.2.4
numpy==1.13.0
oauthlib==2.0.2
pandas==0.20.2
python-dateutil==2.6.0
python-dotenv==0.6.4
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
textblob==0.12.0
tweepy==3.5.0
urllib3==1.21.1
